verificacao dos requisitos

+++ processa.php
@@ -17,7 +17,7 @@ $tipo = $_POST['tipo'];
 $valor = (float) $_POST['valor'];
 $data = $_POST['data'];
 $descricao = $_POST['descricao'];
-$id = count($carteira->getTransacoes()) + 1;
+$id = (int) count($carteira->getTransacoes()) + 1;
 
 try {
     if ($tipo === 'receita') {
